fix: use bfloat16 for MedGemma on Blackwell GPU (fixes precision errors)

+++ src/enso_atlas/reporting/medgemma.py
@@ -133,7 +133,7 @@ class MedGemmaReporter:
 
             # Load model with appropriate precision
             model_kwargs = {
-                "torch_dtype": torch.float16 if self._device.type == "cuda" else torch.float32,
+                "torch_dtype": torch.bfloat16 if self._device.type == "cuda" else torch.float32,
                 "device_map": "auto" if self._device.type == "cuda" else None,
                 "low_cpu_mem_usage": True,
             }
